sentiment_analysis_service/requirements.txt
annotated-types==0.7.0
anyio==4.6.0             
certifi==2024.8.30       
charset-normalizer==3.3.2
click==8.1.7             
colorama==0.4.6          
fastapi==0.115.0         
filelock==3.16.1         
fsspec==2024.9.0
h11==0.14.0
huggingface-hub==0.25.1
idna==3.10
Jinja2==3.1.4
MarkupSafe==3.0.0
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
packaging==24.1
pandas==2.2.3
pydantic==2.9.2
pydantic_core==2.23.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
safetensors==0.4.5
six==1.16.0
sniffio==1.3.1
starlette==0.38.6
sympy==1.13.3
tokenizers==0.20.0
torch==2.4.1
tqdm==4.66.5
transformers==4.45.2
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.31.0